[PLAT-802] Refactor trending
